Fix some capability errors

+++ rts/src/ffi/rts/config.rs
@@ -2,4 +2,4 @@ pub(crate) const RTS_USER_SIGNALS: u32 = 1;
 
 pub(crate) const MAX_N_CAPABILITIES: u32 = 256;
 
-pub(crate) const CACHELINE_SIZE: u32 = 128;
+pub(crate) const CACHELINE_SIZE: usize = 128;
